move utility functions to utils.js, fix test description
